narrow token refresh retry fix
