+++ packages/sandbox/src/util/fresh-auth-retry.ts
@@ -2,7 +2,7 @@ import retry from "async-retry";
 import { APIError } from "@vercel/sandbox";
 import { NotOk } from "@vercel/sandbox/dist/auth/index.js";
 import createDebugger from "debug";
-import { isTokenFresh } from "./auth-freshness";
+import { isTokenFresh } from "../args/auth";
 
 const debug = createDebugger("sandbox:fresh-auth-retry");
 
@@ -41,25 +41,6 @@ function getAuthFailureStatus(error: unknown): number | undefined {
     status = error.response.status;
   } else if (error instanceof NotOk) {
     status = error.response.statusCode;
-  } else if (
-    error &&
-    typeof error === "object" &&
-    "response" in error &&
-    error.response &&
-    typeof error.response === "object"
-  ) {
-    // Errors can cross package/bundle boundaries where instanceof is not
-    // reliable. Both supported auth error shapes expose status on response.
-    const response = error.response as {
-      status?: unknown;
-      statusCode?: unknown;
-    };
-    status =
-      typeof response.status === "number"
-        ? response.status
-        : typeof response.statusCode === "number"
-          ? response.statusCode
-          : undefined;
   }
   return status === 401 || status === 403 ? status : undefined;
 }
